sge: decouple SGE-specific details from grid submission
The upcoming slurm subcommand is going to require a lot of the same
logic as the sge subcommand.  Move those shared parts to grid.go and
introduce a gridSpec struct that captures the scheduler-specific
details.
