+++ docs/validation/matrix.yaml
@@ -64,7 +64,7 @@
   doc: docs/commands/bbi_nonmem_run_sge.md
   tests:
     - cmd/nonmem_test.go
-    - cmd/sge_test.go
+    - cmd/grid_test.go
     - integration/nonmem/bbi_sge_test.go
     - parsers/nmparser/add_path_level_test.go
